+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This release depends on, and has been tested with, the following Mapbox dependen
 - Mapbox Android Core `v5.0.2` ([release notes](https://github.com/mapbox/mapbox-events-android/releases/tag/core-5.0.2))
 
 - Added experimental option to MapboxRouteLineOptions to enable the shared cache in the Maps SDK. [#6680](https://github.com/mapbox/mapbox-navigation-android/pull/6680)
+- Added experimental `MapboxRouteLineOptions#enableSharedCache` option to enable route line's GeoJson source data sharing between multiple instances of the map. [#6680](https://github.com/mapbox/mapbox-navigation-android/pull/6680)
 
 ## Mapbox Navigation SDK 2.10.0-beta.2 - 01 December, 2022
 ### Changelog